continua refactor
